requirements.txt
# 赛博小镇 V2 依赖包
# Cyber Town V2 Requirements

# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# LLM: DeepSeek (via OpenAI-compatible API, no extra package needed)
# Set DEEPSEEK_API_KEY in .env; langchain-openai handles the rest

# PostgreSQL (Long-Term Memory)
asyncpg>=0.29.0

# MCP Protocol
mcp>=1.0.0

# Data Validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0

# Optional: For better async support
aiohttp>=3.9.0

# Optional: Vector Database (if switching back to ChromaDB)
# chromadb>=0.5.0

# Optional: For embedding models
# sentence-transformers>=2.2.0
